taa s

+++ Tutorial 1 Code/Engine/Graphics.h
@@ -78,7 +78,7 @@ public:
 	void Draw_FillTri(int x1, int y1,float z1, int x2, int y2,float z2, int x3, int y3,float z3,float u1,float v1,float u2,float v2,float u3,float v3,UINT* img,Light *light,const class Cmaera  &camera,Vector4d norml);
 	bool Draw_UV(float u1, float v1, float u2, float v2, float u3, float v3, float& u, float& v);
 	void Graphics::Draw_FillTri(int x1, int y1, float z1, int x2, int y2, float z2, int x3, int y3, float z3, float u1, float v1, float u2, float v2, float u3, float v3, UINT* img);
-	bool Draw_UV(float x1, float y1, float x2, float y2, float x3, float y3, float x, float y, float& u, float& v);
+	bool Draw_UV(float x1, float y1, float x2, float y2, float x3, float y3, float x, float y, double& u, double& v);
 	bool Draw_UV(int x1,int y1,int x2,int y2,int x3,int y3,int x,int y,float& u,float& v);
 	void Load2D();
 	BYTE* GetImageData(ID3D11Device* device, ID3D11DeviceContext* context, ID3D11Texture2D* texture, /*OUT*/ int* nWidth, /*OUT*/ int*  nHeight);
